refactor: migrate shared helpers from server.js to helpers.js
